@@ -1,11 +1,23 @@
-import { Component } from '@angular/core';
+import { Component, inject, OnInit } from '@angular/core';
+import { Router, NavigationEnd } from '@angular/router';
 import { IonApp, IonRouterOutlet } from '@ionic/angular/standalone';
+import { MetaService } from './core/services/meta.service';
+import { Meta } from '@angular/platform-browser';
 
 @Component({
   selector: 'app-root',
   templateUrl: 'app.component.html',
   imports: [IonApp, IonRouterOutlet],
 })
-export class AppComponent {
-  constructor() {}
+export class AppComponent implements OnInit {
+  private _meta: Meta = inject(Meta);
+  constructor(private _router: Router, private _metaService: MetaService) {}
+
+  ngOnInit(): void {
+    this._router.events.subscribe((event) => {
+      if (event instanceof NavigationEnd) {
+        this._metaService.updateDescriptionForCurrentRoute(this._meta);
+      }
+    });
+  }
 }
